make graph appear on start
