+++ utils/data/starting-input.js
@@ -1,0 +1,29 @@
+const startingInput = {
+  total: {
+    valueSum: 0,
+    valueSumIst: 0,
+    difference: 0,
+  },
+
+  "9b51189fa12": {
+    value: 0,
+    valueIst: 0,
+  },
+  b51189fa126: {
+    value: 0,
+    valueIst: 0,
+  },
+  "51189fa126b": {
+    value: 0,
+    valueIst: 0,
+  },
+  "1189fa126b6": {
+    value: 0,
+    valueIst: 0,
+  },
+  be90e393b31: {
+    value: 0,
+    valueIst: 0,
+  },
+};
+export default startingInput;
